Improves video ad demo and emulator support
Refactors the video ad demo screen to include all possible delivery, end card and skippable combinations.

Adds functionality to rewrite localhost URLs to 10.0.2.2 for compatibility with the Android emulator.
This ensures that video ads and tracking beacons work correctly when the app is running in the emulator.

Also, adjusts mock server URL to http from https.

Relates to 288-create-example-app-flow-video